Cast the string lengths that narrow on x64
size_t is 64-bit on x64, so every `len = strlen(...)` into a UDWORD,
SDWORD or int is a C4267. 31 sites across ten files, all of them
lengths of names, console messages, edit-box contents and file paths
that cannot approach 2GB, so this is noise removal rather than a fix --
but it is the noise that buries the findings that matter.

The cast goes to whatever the variable is already declared as, so the
arithmetic is unchanged. The GfxFile sites in Structure.cpp keep doing
the `- 5` in size_t before the cast, which is what they did as UDWORD
before.

+++ NeuronClient/TextDraw.cpp
@@ -353,7 +353,7 @@ void SetExtentsMode_USEMAX(void) { ExtentsMode = EXTENTS_USEMAXWIDTH; }
 UDWORD pie_DrawFormattedText(UBYTE* String, UDWORD x, UDWORD y, UDWORD Width, UDWORD Justify, BOOL DrawBack)
 {
   int i, si, osi;
-  int Len = strlen((char*)String);
+  int Len = static_cast<int>(strlen((char*)String));
   int jx = x; // Default to left justify.
   int jy = y;
   int WWidth;
